third_party/blink/web_tests/flag-specific/highdpi/virtual/composite-after-paint/scrollingcoordinator/non-fast-scrollable-region-nested-expected.txt

The test ensures that non-fast scrollable regions nested in iframe are correctly offset by the iframe location.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS nonFastScrollableRects.length is 3
FAIL rectToString(nonFastScrollableRects[0]) should be [51, 102, 200, 200]. Was [77, 153, 300, 300].
FAIL rectToString(nonFastScrollableRects[1]) should be [51, 402, 211, 211]. Was [77, 603, 317, 317].
FAIL rectToString(nonFastScrollableRects[2]) should be [51, 702, 222, 222]. Was [77, 1053, 333, 333].
PASS successfullyParsed is true

TEST COMPLETE
